+++ src/search/landmarks/landmark_factory_rpg_exhaust.cc
@@ -23,7 +23,7 @@ void LandmarkFactoryRpgExhaust::generate_landmarks() {
     }
     // test all other possible facts
     const State &initial_state = g_initial_state();
-    for (int i = 0; i < g_variable_name.size(); ++i)
+    for (size_t i = 0; i < g_variable_name.size(); ++i)
         for (int j = 0; j < g_variable_domain[i]; ++j) {
             const pair<int, int> lm = make_pair(i, j);
             if (!lm_graph->simple_landmark_exists(lm)) {
